Move DAL registration to BLL and cleanup
