+++ MiniEcommerce.Contracts/Interfaces/IUnitOfWork.cs
@@ -6,9 +6,8 @@ using System.Text;
 namespace MiniEcommerce.Contracts.Interfaces;
 
 public interface IUnitOfWork
-{
+{   
     IProductRepository Products { get; }
     IRepository<Category> Categories { get; }
-
     Task<int> SaveChangesAsync(CancellationToken cancellationToken);
 }
