+++ MiniEcommerce/Program.cs
@@ -1,6 +1,5 @@
 using Microsoft.EntityFrameworkCore;
-using MiniEcommerce.DataAccessLayer.Extensions;
-using MiniEcommerce.BusinessLogicLayer.Extentions;
+using MiniEcommerce.BusinessLogicLayer.Extensions;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -9,8 +8,7 @@ var builder = WebApplication.CreateBuilder(args);
 builder.Services.AddControllers();
 // Learn more about configuring OpenAPI at https://aka.ms/aspnet/openapi
 builder.Services.AddOpenApi();
-builder.Services.AddDataAccessLayer(builder.Configuration);
-builder.Services.AddMapping();
+builder.Services.AddBusinessLogicLayer(builder.Configuration);
 
 var app = builder.Build();
